+++ airflow/config_templates/default_airflow.cfg
@@ -704,7 +704,7 @@ password =
 # provided explicitly or passed via ``default_args``
 default_owner = airflow
 
-# The default value of attribute "deferrable" in BaseDeferrableOperator.
+# The default value of attribute "deferrable" in DeferrableMixin.
 default_deferrable = false
 default_cpus = 1
 default_ram = 512
