+++ airflow/models/basedeferrableoperator.py
@@ -1,0 +1,39 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+from __future__ import annotations
+
+from airflow.configuration import conf
+from airflow.models.baseoperator import BaseOperator
+
+DEFAULT_DEFERRABLE: bool = conf.getboolean("operators", "default_deferrable")
+
+
+class BaseDeferrableOperator(BaseOperator):
+    """Abstract base class for deferrable operators.
+
+    It comes with "deferrable" attribute which should
+    be used in the execute method to decide whether to
+    run this task in deferrable mode.
+
+    One can configurate all tasks inherit this BaseDeferrableOperator
+    through DEFAULT_DEFERRABLE which can be configurated through airflow.cfg
+    """
+
+    def __init__(self, *args, deferrable=DEFAULT_DEFERRABLE, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.deferrable = deferrable
