+++ airflow/providers/amazon/aws/hooks/base_aws.py
@@ -63,7 +63,7 @@ from airflow.utils.log.secrets_masker import mask_secret
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
