+++ airflow/providers/amazon/aws/operators/glue.py
@@ -33,7 +33,7 @@ from airflow.providers.amazon.aws.triggers.glue import GlueJobCompleteTrigger
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
