+++ airflow/providers/amazon/aws/operators/glue_crawler.py
@@ -27,7 +27,7 @@ from airflow.providers.amazon.aws.triggers.glue_crawler import GlueCrawlerComple
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
