+++ airflow/providers/amazon/aws/operators/redshift_cluster.py
@@ -35,7 +35,7 @@ from airflow.providers.amazon.aws.triggers.redshift_cluster import (
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
