+++ airflow/providers/amazon/aws/operators/sagemaker.py
@@ -39,7 +39,7 @@ from airflow.utils.json import AirflowJsonEncoder
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
 if TYPE_CHECKING:
