+++ airflow/providers/databricks/operators/databricks.py
@@ -34,7 +34,7 @@ from airflow.providers.databricks.utils.databricks import normalise_json_content
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
