+++ airflow/providers/dbt/cloud/operators/dbt.py
@@ -36,7 +36,7 @@ from airflow.providers.dbt.cloud.triggers.dbt import DbtCloudRunJobTrigger
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
