+++ airflow/providers/dbt/cloud/sensors/dbt.py
@@ -29,7 +29,7 @@ from airflow.sensors.base import BaseSensorOperator
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
