+++ airflow/providers/google/cloud/operators/bigquery.py
@@ -57,7 +57,7 @@ from airflow.providers.google.cloud.utils.bigquery import convert_job_id
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
