+++ airflow/providers/google/cloud/operators/bigquery_dts.py
@@ -41,7 +41,7 @@ from airflow.providers.google.cloud.triggers.bigquery_dts import BigQueryDataTra
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
