+++ airflow/providers/google/cloud/operators/cloud_build.py
@@ -46,7 +46,7 @@ from airflow.utils.helpers import exactly_one
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
