+++ airflow/providers/google/cloud/operators/dataflow.py
@@ -45,7 +45,7 @@ from airflow.version import version
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
