+++ airflow/providers/google/cloud/operators/kubernetes_engine.py
@@ -40,7 +40,7 @@ from airflow.utils.timezone import utcnow
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
