+++ airflow/providers/google/cloud/operators/mlengine.py
@@ -43,7 +43,7 @@ from airflow.providers.google.cloud.triggers.mlengine import MLEngineStartTraini
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
