+++ airflow/providers/google/cloud/sensors/gcs.py
@@ -41,7 +41,7 @@ from airflow.sensors.base import BaseSensorOperator, poke_mode_only
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
