+++ airflow/providers/google/cloud/transfers/bigquery_to_gcs.py
@@ -34,7 +34,7 @@ from airflow.providers.google.cloud.triggers.bigquery import BigQueryInsertJobTr
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
