+++ airflow/providers/google/cloud/transfers/gcs_to_bigquery.py
@@ -46,7 +46,7 @@ from airflow.providers.google.cloud.triggers.bigquery import BigQueryInsertJobTr
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
 if TYPE_CHECKING:
