+++ airflow/providers/microsoft/azure/operators/data_factory.py
@@ -37,7 +37,7 @@ from airflow.utils.log.logging_mixin import LoggingMixin
 # remove this try-exception block and
 # inherit BaseDeferrableOperator for operator with deferrable attribute
 try:
-    from airflow.models.baseoperator import DEFAULT_DEFERRABLE
+    from airflow.models.basedeferrableoperator import DEFAULT_DEFERRABLE
 except ImportError:
     DEFAULT_DEFERRABLE = False
